--- src/main/java/util/persistence/sql/ddl/PersonV0.java
@@ -1,4 +1,0 @@
-package util.persistence.sql.ddl;
-
-public class PersonV0 {
-}
