+++ src/test/java/persistence/sql/ddl/PersonV0.java
@@ -1,0 +1,4 @@
+package persistence.sql.ddl;
+
+public class PersonV0 {
+}
