+++ src/test/java/persistence/sql/ddl/PersonV1.java
@@ -1,4 +1,4 @@
-package util.persistence.sql.ddl;
+package persistence.sql.ddl;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
