+++ src/test/java/persistence/sql/mapping/ColumnTypeMapperTest.java
@@ -1,4 +1,4 @@
-package util.persistence.sql.mapping;
+package persistence.sql.mapping;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
