+++ src/test/java/persistence/study/ReflectionTest.java
@@ -1,4 +1,4 @@
-package util.persistence.study;
+package persistence.study;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
